pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rudebench"
version = "0.2.0"
description = "Multi-dimensional behavioral benchmark measuring how LLMs change behavior under hostile prompting"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.40.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
rudebench = "rudebench.__main__:main"

[tool.setuptools.packages.find]
include = ["rudebench*"]
